Fix options for collectiondbk2pdf in build-pdf
According to README, it is possible to pass in the path to prince:

```
./script/build-pdf ${book_name} -p ${path-to-princexml-binary}
```

The script was not passing the path to princexml binary correctly to
collectiondbk2pdf.  It was only passing the third argument to _build_pdf
and the second argument was used as the xhtml filename.

This change adds a usage message:

```
Usage: ./script/build-pdf <book-name> [<xhtml_filename>] [-p <pdfgen>] [-r]

Optional arguments:
  -p [PDFGEN]        Path to a PDF generation script
  -r                 Reduce image quality
```

and fixes the code that parses the arguments.